requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
django>=5.0.0
psycopg2-binary>=2.9.9
asgiref>=3.8.0
firebase-admin>=6.5.0
anthropic>=0.30.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
django-cors-headers>=4.4.0
